counter_app/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
gunicorn==21.2.0
prometheus-flask-exporter==0.23.0